Use WebApplicationFactory HTTP transport
